refactor(core): extract nearest-nav helpers from candidate-geometry (#534)
Move closestOrthInRange, directionalNearestInOrder, and panelRangeInOrder
into candidate-geometry-nearest.ts. Re-export from candidate-geometry for
existing imports; eager store imports the dedicated module.
